requirements.txt
flask
feedparser
requests
python-dotenv
